requirements.txt
absl-py
alabaster
annotated-types
anyio
arabic-reshaper
argon2-cffi
argon2-cffi-bindings
arrow
asgiref
asn1crypto
asttokens
astunparse
async-lru
attr
attrs
Autils
babel
backports.tarfile
beautifulsoup4
bleach
blis
Brotli
brotlicffi
cachetools
catalogue
certifi
cffi
chardet
charset-normalizer
click
cloudpathlib
colorama
comm
confection
configparser
contourpy
cryptography
cssselect2
cycler
cymem
debugpy
decorator
defusedxml
Django
django-widget-tweaks
docutils
docx2pdf
en_core_web_sm @ https://github.com/explosion/spacy-models/releases/download/en_core_web_sm-3.8.0/en_core_web_sm-3.8.0-py3-none-any.whl#sha256=1932429db727d4bff3deed6b34cfc05df17794f4a52eeb26cf8928f7c1a0fb85
executing
fanstatic
fastjsonschema
filelock
flatbuffers
fonttools
fqdn
fsspec
gast
google-api-core
google-api-python-client
google-auth
google-auth-httplib2
google-auth-oauthlib
google-pasta
googleapis-common-protos
grpcio
h11
h2
h5py
hpack
html5lib
HTMLParser
httpcore
httplib2
httpx
huggingface-hub
hyperframe
idna
imagesize
importlib_metadata
ipykernel
ipython
ipywidgets
isoduration
jaraco.classes
jaraco.context
jaraco.functools
jedi
Jinja2
joblib
js
json5
jsonpointer
jsonschema
jsonschema-specifications
jupyter-events
jupyter-lsp
jupyter_client
jupyter_core
jupyter_server
jupyter_server_terminals
jupyterlab
jupyterlab_pygments
jupyterlab_server
jupyterlab_widgets
keras
keyring
kiwisolver
langcodes
language_data
libclang
lightgbm
lxml
marisa-trie
Markdown
markdown-it-py
MarkupSafe
matplotlib
matplotlib-inline
mdurl
mistune
ml-dtypes
more-itertools
mpmath
murmurhash
namex
nbclient
nbconvert
nbformat
nest-asyncio
networkx
notebook
notebook_shim
numpy
oauthlib
opencv-python
opt_einsum
optree
oscrypto
overrides
packaging
pandas
pandocfilters
parso
patsy
Pillow
platformdirs
preshed
prometheus_client
prompt_toolkit
proto-plus
protobuf
psutil
pure_eval
pyasn1
pyasn1_modules
pycparser
pydantic
pydantic_core
Pygments
pyHanko
pyhanko-certvalidator
pyOpenSSL
pyotp
pyparsing
pypdf
pypng
python-bidi
python-dateutil
python-docx
python-json-logger
pytz
pywin32
pywin32-ctypes
pywinpty
PyYAML
pyzmq
qrcode
redis
referencing
regex
reportlab
requests
requests-oauthlib
rfc3339-validator
rfc3986-validator
rich
roman-numerals-py
rpds-py
rsa
safetensors
scikit-learn
scipy
seaborn
Send2Trash
sentence-transformers
shellingham
simplejson
six
smart_open
sniffio
snowballstemmer
sns
soupsieve
spacy
spacy-legacy
spacy-loggers
Sphinx
sphinxcontrib-applehelp
sphinxcontrib-devhelp
sphinxcontrib-htmlhelp
sphinxcontrib-jsmath
sphinxcontrib-qthelp
sphinxcontrib-serializinghtml
sqlparse
srsly
stack-data
statsmodels
svglib
sympy
tensorboard
tensorboard-data-server
tensorflow
tensorflow-io-gcs-filesystem
tensorflow_intel
termcolor
terminado
tf_keras
thinc
thread
threadpoolctl
tinycss2
tokenizers
torch
tornado
tqdm
traitlets
transformers
typer
types-python-dateutil
typing-inspection
typing_extensions
tzdata
tzlocal
uri-template
uritemplate
uritools
urllib3
urllib3-secure-extra
wasabi
wcwidth
weasel
webcolors
webencodings
WebOb
websocket-client
Werkzeug
widgetsnbextension
wrapt
xhtml2pdf
zipp
zstandard
